OKAY ITS CLEANED UP
